requirements.txt
django~=5.0.6
djangorestframework~=3.15.1
whitenoise
python-dotenv~=1.0.1
crispy-bootstrap5
Pillow
psycopg2-binary
python-telegram-bot~=20.3
requests~=2.32.2
beautifulsoup4